requirements.txt
aspose-words==22.4.0
click==8.1.3
lxml==4.8.0
numpy==1.22.3
opencv-python==4.5.5.64
packaging==21.3
pdf2image==1.16.0
Pillow==9.1.0
poppler-utils==0.1.0
PyMuPDF==1.19.6
pyparsing==3.0.8
PyPDF2==1.27.12
pytesseract==0.3.9
python-docx==0.8.11
